Updates to installation scripts and files
Testing discovered that the upgrade process fails when `dpkg` installs files directly to the ESP, because you can't create symlinks to a FAT32 filesystem. Set up the package to install to `/usr/lib` instead, and then copy the files to their working locations via `postinst` script, and remove them via `postrm` script. So far, so good.