Added `memtest-util.sh` script, adjusted package installer to use
Per conversation with others, I've added a `memtest-util.sh` script with `uninstall` and `reinstall` parameters. This script does the actual work of copying/removing the memtest files from the ESP, and can be run by a user after package installation. This script is also now called by the `postinst` and `prerm` package scripts, and generates the `memtest.conf` file instead of just copying it.